Reorganize codebase with file renaming, better documentation, and bug fixes
This commit includes renaming KronotopInstanceTest to KronotopTestInstance. New classes were also added, like KronotopTestCluster and BeforeAfterSuiteListener. Several files across the codebase have more informative comments now, making it easier to understand the purpose and functionality of the classes and methods. The FoundationDB version was updated in the pom.xml file which improves the working of the application and ensures consistency across different setups. Additionally, smaller bugs and inconsistencies were fixed in various classes.