Replace 'resolveKey' method usage with 'findShard'
Refactored the code to use 'findShard' method instead of 'resolveKey' in all classes. The refactor improves clarity as 'findShard' more appropriately communicates the method's function of locating the Shard associated with a key or set of keys. The change applies to several classes in the 'redis' package and 'RedisService' where the method definitions were updated.