Add GitHub Actions workflow and update FoundationDB config
Added a new GitHub Actions workflow for Maven and Java CI. This workflow runs on push or pull request to the main branch and tests the project against Java versions 11, 17, and 21. It also downloads and installs FoundationDB server and client before running the tests. Minor updates were made to the FoundationDBFactory class to accommodate this change. This not only streamlines the CI/CD process but also ensures the project is compatible across multiple Java versions.