Extend the Response interface and implementations in server package
The `Response` interface and its implementations `TransactionResponse` and `RespResponse` have been expanded in this commit. Methods to write double values, maps, null values and booleans were added. Also, the `writeMap()` method is utilised in the `HelloHandler` class. It modifies `GetReadVersionHandler` and corresponding test to write readVersion as double instead of integer.