Implement sqlGetSchema functionality
Added support for sqlGetSchema operation in the existing code. This includes creating new handlers, SQLGetSchemaHandler and SqlGetSchemaMessage and adding SQLGETSCHEMA in the MessageTypes. This update enhances the SQL service capabilities of the system by allowing users to fetch the current SQL schema information. Necessary adjustments were also made in the test files to accommodate these changes.