Add support for 'DROP COLUMN' in ALTER TABLE
This commit adds functionality to support the 'DROP COLUMN' operation in the ALTER TABLE command. The feature was achieved by adding new classes like DropColumn and ColumnNotExistsException, modifying existing classes like AlterTable and AddColumn, and updating tests to cover the new functionality. In addition, some refactoring was done to streamline the code such as removing the SqlValidationException class and renaming it to SqlExecutionException.