Update SqlMetadataService and enhance schema handling tests
The checkAndLoadSchema method in SqlMetadataService is now public, improving its accessibility. Adjustments were also made to the SET schema test in SetOptionTest file. Specifically, a new test was added for handling a non-existing schema. Modifications were also made in SetOption.java to handle absence of a schema and existence of an invalid schema.