Add 'get' method in 'Volume' and update dependencies
A new method 'get' has been added to the 'Volume' class for retrieving data based on the provided key. The 'netty' dependency versions have been updated in 'pom.xml'. Additional cleanup and minor changes are also included, such as an unnecessary context in the 'Segment' class and a new 'TODO' comment. A call to the newly added 'get' method has also been included in the 'VolumeTest'.