Implement caching for EntryMetadata in Volume class
Caching for EntryMetadata has been introduced in the Volume class using Guava's LoadingCache. This cache is built in the Volume constructor and is used to cache EntryMetadata within the append method and retrieve it in the get method. The test class has been updated to demonstrate this caching behavior. The EntryMetadata encode method has also been rearranged, but its functionality remains the same.