Refactor 'get' method in Volume class
The 'get' method has been refactored for better separation of concerns. The cache loading functionality was moved into a new private method 'loadEntryMetadataFromCache'. This enhances readability and maintainability of the 'get' method. In case the 'SegmentNotFoundException' occurs, the cache is invalidated and reloaded from FoundationDB, accounting for possible segment cleanup.