Add flush method and update ByteBuffer handling
The flush method has been added to the volume and segment classes to persist changes. Additionally, the handling of the ByteBuffer in the encode() and decode() methods of the SegmentMetadata class has been improved with the adding of the flip() method after the final put operation. The get() method in the segment class has also been refactored and simplified, removing unnecessary locking. A test to handle reopening of the volume has been added as well.