Add getEntries helper method in VolumeTest
The commit introduces a helper method, getEntries, to reduce redundancy in the test codebase. This method streamlines the creation of ByteBuffer arrays by generating the 'entries' data used in the append, get, and delete tests in VolumeTest. It uses the Google Guava's Strings class to pad the start of each entry.