Add user version implementation and segment analysis
The main change in this commit is adding the UserVersion.java class to manage and increment a user version. Furthermore, a new segment analysis method is introduced in our Volume.java. This method checks and calculates the used bytes and cardinality of each segment in the volume. Additionally, the introduction of the Vacuum.java class prepares for upcoming features. Lastly, relevant unit tests are added to ensure the correctness of the new functionalities.