Refactor Volume class to use Session instead of Transaction directly
This update removes the `UserVersion` field in the `Volume` class and instead uses `Session` to manage the user version. It adds a check in `append()` to limit the number of entries and throws a new `TooManyEntriesException` when this limit is exceeded. All methods and tests that were using `Transaction` have been updated to use `Session` instead.