Implement vacuum feature in volume module
The vacuum feature has been added to the volume module. It includes the analyze method which is now set as protected and a newly introduced evictSegment method. Also, the VolumeConfig class has been adapted to include an allowed garbage ratio. Lastly, the Vacuum class has been augmented with the ability to reset itself and to perform vacuum operations.