Replace ConcurrentHashMap with HashMap in Volume class
The ConcurrentHashMap of segmentLogs in the Volume class was replaced with a standard HashMap to better manage data access. Additionally, a segmentsLock was added as a read lock for better concurrency control. Also, an exception was included to handle the case where a segment queried in appendSegmentLog does not exist.