Refactor test configuration and setup
Refactored the BaseTest and BaseClusterTest classes to simplify and standardize the test configuration process. Java Paths library is now used for setting the root path in the loadConfig method instead of absolute paths. Tests now use a constant cluster name and a unique volume root path for every invocation of the loadConfig method.