Replace SlotHash with JedisClusterCRC16 and add jedis dependency
The implementation of key slot calculation has been switched from SlotHash to JedisClusterCRC16. A new file, InternalCommand.java, was added to support this change. Also, the jedis library was added as a dependency in pom.xml to support Redis-specific operations. The imports in several files were updated to use JedisClusterCRC16 instead of SlotHash.