Add replication control and data segmentation methods
Refactored the Replication class in the volume package to enhance control of replication processes. New start and stop methods are introduced for the application control flow. Data segmentation logic has also been updated with the addition of private methods to fetch and insert segment range which increase data processing efficiency.