Refactor replication code and add JSONUtils import
The replication methods insertSegmentRange and fetchSegmentRange were rearranged in the code to enhance readability. The start method was updated to include setting isStarted as true. For the SnapshotJob subclass, the code handling VersionstampedKeySelector was updated for better handling of snapshots, and result handling was optimized. Import for JSONUtils was also added for future utilization.