Refactor logic related to snapshot job completion
The logger in SnapshotJobRunner has been corrected to reference its own class. Changes were made in the isSnapshotCompleted() method to factor out the calculation of job completion from the metadata calculation process. Additional logging of the number of processed keys once the snapshot job is complete is also included for better tracking and visibility.