Add condition to check replication job status before starting
The update adds conditions in the Replication class to check if replication jobs are complete before starting. This also involved the expansion of the start() method and creation of a new method, isSnapshotCompleted(). Now, there are logging messages indicating the creation, start, and stop of replication jobs and snapshot completion are added for better