Add CDC-only mode to replication process
This commit introduces a new configuration option, `cdcOnly`, to the replication process, allowing it to bypass the snapshot stage if desired. Adjustments are made in the `Replication` class to check for the `cdcOnly` flag and alter the flow accordingly. Additionally, a constructor change in `ReplicationConfig` and an update in the `ReplicationTest` unit test reflect the inclusion of this new option.